extrator de pdf/requirements.txt
streamlit==1.37.1
numpy==1.26.4
pandas==2.2.2
plotly==5.24.1
Pillow==10.2.0
PyMuPDF==1.23.26
openpyxl==3.1.2 